add real counter to user counter component on buy credit increase size of response menu improve generate for reply language Add scroll to top to footer add admin route for user counter
