+++ pages/api/admin/user.ts
@@ -1,0 +1,31 @@
+import { NextApiRequest, NextApiResponse } from 'next'
+import { errorResponse, successResponse } from '@/lib/helpers'
+import connectToMongoDb from '@/lib/connectDb'
+import { User } from 'models/User'
+
+const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const { method } = req
+  switch (method) {
+    case 'GET':
+      try {
+        await connectToMongoDb()
+        const users = await User.find({})
+
+        return successResponse(
+          res,
+          'user profiles fetched successfully',
+          { users_count: users.length },
+          200,
+        )
+      } catch (error) {
+        return errorResponse(res, 'something went wrong', 400)
+      }
+
+    default:
+      res.setHeader('Allow', ['GET'])
+      res.status(405).end(`Method ${method} Not Allowed`)
+      break
+  }
+}
+
+export default handler
